fix(jest-config): make frontend testing work in standalone extensions
The documented frontend test setup only worked inside the monorepo. A
standalone, Composer-installed extension could not run it: @flarum/core
is not an npm package, so the bootstrap imports were unresolvable, and
even mapped to the vendored core its TypeScript went untransformed and
its runtime deps were missing.

- Resolve @flarum/core to the Composer-vendored core (vendor/flarum/core/js)
  when present, falling back to the monorepo package. Core and bundled
  extensions are unaffected.
- Add core's vendored source to the transform roots so its TypeScript is
  compiled, and the extension's node_modules to the resolver paths so
  core's own imports (mithril, format-message, ...) resolve there from a
  single `yarn install` in the extension's js folder.
- Read core's locale file from the resolved core dir rather than a bare
  relative path.
- Move the frontend test factory (makeUser/makeDiscussion) into this
  package, which extensions already install, instead of core's tests/
  directory, which the Composer package strips. Core re-exports it so
  existing monorepo imports keep working. No test code ships to
  production installs.
- Carry core's frontend runtime dependencies so the single install
  provides everything core's source needs at test time.
