+++ js-packages/jest-config/src/bootstrap/common.js
@@ -1,8 +1,9 @@
 import Drawer from '@flarum/core/src/common/utils/Drawer';
-import { makeUser } from '@flarum/core/tests/factory';
+import { makeUser } from '../../factory';
 import flatten from 'flat';
 import jsYaml from 'js-yaml';
 import fs from 'fs';
+import path from 'path';
 
 export default function bootstrap(Application, app, payload = {}) {
   Application.prototype.mount = () => {};
@@ -37,6 +38,9 @@ export default function bootstrap(Application, app, payload = {}) {
   });
 
   app.translator.setLocale('en');
-  app.translator.addTranslations(flatten(jsYaml.load(fs.readFileSync('../locale/core.yml', 'utf8'))));
+  // core's locale file lives at <core>/locale/core.yml. `__FLARUM_CORE_DIR__`
+  // points at core's `js` dir (set by the jest config), so step up one level.
+  const coreLocale = path.join(__FLARUM_CORE_DIR__, '..', 'locale', 'core.yml');
+  app.translator.addTranslations(flatten(jsYaml.load(fs.readFileSync(coreLocale, 'utf8'))));
   app.drawer = new Drawer();
 }
